# services/api-gateway/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
...
  handlers_admin.updateClinicScheduleRequest:
    properties:
      schedule:
        items:
          $ref: '#/definitions/model.ClinicWeeklySchedule'
        type: array
      slot_duration_minutes:
        type: integer
    type: object
  handlers_admin.updateDoctorScheduleRequest:
    properties:
      schedule:
        items:
          $ref: '#/definitions/model.DoctorWeeklySchedule'
        type: array
      slot_minutes:
        type: integer
    type: object
...
  model.AdminWithRole:
    properties:
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      phone:
        type: string
      role:
        type: string
      secondName:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
  model.ClinicDailyOverride:
    properties:
      date:
        type: string
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      type:
        type: string
    type: object
  model.ClinicWeeklySchedule:
    properties:
      day:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      is_day_off:
        type: boolean
      slot_minutes:
        type: integer
      start_time:
        type: string
    type: object
...
  model.DoctorDailyOverride:
    properties:
      date:
        type: string
      doctor_id:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      start_time:
        type: string
      type:
        type: string
    type: object
...
  model.DoctorWeeklySchedule:
    properties:
      day:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      is_day_off:
        type: boolean
      selectedDoctor:
        type: integer
      slot_minutes:
        type: integer
      start_time:
        type: string
    type: object
...
  model.Patient:
    properties:
      birthDate:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      password:
        type: string
      phone:
        type: string
      secondName:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
  model.PatientWithoutPassword:
    properties:
      birthDate:
        type: string
      email:
        type: string
      firstName:
        type: string
      gender:
        type: string
      phone:
        type: string
      secondName:
        type: string
      surname:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: ПО для стоматологической клиники
  title: Примастом API-Gateway
  version: "1.0"
paths:
...
  /api/admin/me:
    get:
      description: Возвращает информацию о текущем администраторе по токену
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AdminWithRole'
        "401":
          description: Необходима авторизация
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Получить профиль администратора
      tags:
      - Администратор
...
  /api/clinic-overrides:
    post:
      consumes:
      - application/json
      description: Устанавливает новое расписание или выходной день для клиники на
        указанную дату
      parameters:
      - description: Переопределение дня
        in: body
        name: override
        required: true
        schema:
          $ref: '#/definitions/model.ClinicDailyOverride'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Добавить переопределение расписания клиники на конкретный день
      tags:
      - Администратор
  /api/clinic-schedule:
    put:
      consumes:
      - application/json
      description: Заменяет расписание клиники на новое, с заданной продолжительностью
        слота
      parameters:
      - description: Новое расписание клиники
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/handlers_admin.updateClinicScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить недельное расписание клиники
      tags:
      - Администратор
...
  /api/doctor-overrides:
    post:
      consumes:
      - application/json
      description: Устанавливает новое расписание или выходной день для врача на указанную
        дату
      parameters:
      - description: Переопределение дня врача
        in: body
        name: override
        required: true
        schema:
          $ref: '#/definitions/model.DoctorDailyOverride'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Добавить переопределение расписания врача на конкретный день
      tags:
      - Администратор
  /api/doctor-schedule/{selectedDoctor}:
    put:
      consumes:
      - application/json
      description: Заменяет расписание врача на новое, с заданной продолжительностью
        слота
      parameters:
      - description: ID врача
        in: path
        name: selectedDoctor
        required: true
        type: integer
      - description: Новое расписание врача
        in: body
        name: schedule
        required: true
        schema:
          $ref: '#/definitions/handlers_admin.updateDoctorScheduleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Некорректные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить недельное расписание врача
      tags:
      - Администратор
...
  /api/patients:
    get:
      description: Возвращает всех зарегистрированных пациентов
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.PatientWithoutPassword'
            type: array
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Получить список пациентов
      tags:
      - Администратор
...
  /api/patients/{id}:
    delete:
      description: Удаляет пользователя по ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверный ввод
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Удалить пользователя
      tags:
      - Администратор
    put:
      consumes:
      - application/json
      description: Обновляет информацию о пациенте по ID
      parameters:
      - description: ID пациента
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные пациента
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/model.Patient'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/gin.H'
        "403":
          description: Недостаточно прав
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/gin.H'
      summary: Обновить данные пациента
      tags:
      - Администратор
...
swagger: "2.0"